added clang nvcc gcc icpc icpx, seperated into float(fp32) and integer

+++ float/nvcc/source_gpu_optimized.cu
@@ -2,25 +2,30 @@
     #include <math.h>
     #include <chrono>
     #include <algorithm>
+    #include "cuda.h"
+    #include "cuda_fp16.h"
+    
 
         using namespace std;
         using namespace std::chrono;
 
 
-
     // CUDA kernel to add elements of two arrays
  __global__
-    void add(int n, float *x, float *y)
+    void add(int n, float*x, float*y)
     {
       int index = blockIdx.x * blockDim.x + threadIdx.x;
       int stride = blockDim.x * gridDim.x;
-      for (int i = index; i < n; i += stride)
-        y[i] = x[i] + y[i];
+//	 for (int i = index; i < n; i += stride)
+//	y[i] = __device__ ​ __half __hadd ( const __half x[i], const __half y[i] ) ;  
+
+	 for (int i = index; i < n; i += stride)
+		 y[i] = x[i] + y[i];
     }
 
 
 __global__
-	void init(int n, float *x, float *y) 
+	void init(int n, float*x, float*y) 
 {
   	int index = threadIdx.x + blockIdx.x * blockDim.x;
   	int stride = blockDim.x * gridDim.x;
@@ -42,11 +47,11 @@ __global__
 	std::chrono::time_point<std::chrono::system_clock> start,stop;
         start = std::chrono::system_clock::now();;
 
-	int count = 64;
+	int count =64;
 
 	while (count >0){
       int N = 1<<27;
-      float *x, *y;
+	 float *x, *y;
      
       // Allocate Unified Memory -- accessible from CPU or GPU
       cudaMallocManaged(&x, N*sizeof(float));
@@ -70,7 +75,7 @@ __global__
      
       // Wait for GPU to finish before accessing on host
 	cudaDeviceSynchronize();
-     
+
       // Check for errors (all values should be 3.0f)
       float maxError = 0.0f;
       for (int i = 0; i < N; i++)
@@ -84,7 +89,7 @@ __global__
 	}
      // calculate and output time 
 	stop = std::chrono::system_clock::now();
-        std::chrono::duration<double> elapsed_seconds = stop - start;
+        std::chrono::duration<float> elapsed_seconds = stop - start;
 
 //	std::cout << "Max error: " << maxError << std::endl;
