+++ integer/nvcc/source_gpu_baseline_integer.cu
@@ -12,7 +12,7 @@
 
     // CUDA kernel to add elements of two arrays
     __global__
-    void add(int n, float *x, float *y)
+    void add(int n, int *x, int *y)
     {
       int index = blockIdx.x * blockDim.x + threadIdx.x;
       int stride = blockDim.x * gridDim.x;
@@ -33,11 +33,11 @@
 	while (count > 0){ 
 
 	int N = 1<<27;
-      	float *x, *y;
+      	int *x, *y;
      
       // Allocate Unified Memory -- accessible from CPU or GPU
-      cudaMallocManaged(&x, N*sizeof(float));
-      cudaMallocManaged(&y, N*sizeof(float));
+      cudaMallocManaged(&x, N*sizeof(int));
+      cudaMallocManaged(&y, N*sizeof(int));
      
       // initialize x and y arrays on the host
       for (int i = 0; i < N; i++) {
@@ -53,13 +53,13 @@
       // Wait for GPU to finish before accessing on host
       cudaDeviceSynchronize();
 
-      // Check for errors (all values should be 3.0f)
+/*      // Check for errors (all values should be 3.0f)
       float maxError = 0.0f;
       for (int i = 0; i < N; i++)
         maxError = fmax(maxError, fabs(y[i]-3.0f));
 
 	if(count<2){	std::cout << "Max error: " << maxError << std::endl;}
-
+*/
       // Free memory
       cudaFree(x);
       cudaFree(y);
